# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mawaeedak API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unread_only
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /daily-messages:
    get:
      operationId: listDailyMessages
      tags: [messages]
      summary: List daily messages
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyMessage"
    post:
      operationId: createDailyMessage
      tags: [messages]
      summary: Create daily message (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyMessage"
...
  /daily-messages/{id}:
    patch:
      operationId: updateDailyMessage
      tags: [messages]
      summary: Update daily message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyMessageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyMessage"
    delete:
      operationId: deleteDailyMessage
      tags: [messages]
      summary: Delete daily message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Themes
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: News
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Jobs
...
  /public-events:
    get:
      operationId: listPublicEvents
      tags: [public-events]
      summary: List public events
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Public events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicEvent"
    post:
      operationId: createPublicEvent
      tags: [public-events]
      summary: Create public event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEvent"
  /public-events/{id}:
    patch:
      operationId: updatePublicEvent
      tags: [public-events]
      summary: Update public event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEvent"
    delete:
      operationId: deletePublicEvent
      tags: [public-events]
      summary: Delete public event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Complaints
...
components:
  schemas:
...
    Notification:
      type: object
      required: [id, title, type, is_read, created_at]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: ["string", "null"] }
        type: { type: string }
        is_read: { type: boolean }
        created_at: { type: string }
    NotificationInput:
      type: object
      required: [title, type]
      properties:
        title: { type: string }
        body: { type: string }
        type: { type: string }
    DailyMessage:
      type: object
      required: [id, message, is_active, created_at]
      properties:
        id: { type: integer }
        message: { type: string }
        display_date: { type: ["string", "null"] }
        is_active: { type: boolean }
        created_at: { type: string }
    DailyMessageInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
        display_date: { type: string }
        is_active: { type: boolean }
    DailyMessageUpdate:
      type: object
      properties:
        message: { type: string }
        display_date: { type: string }
        is_active: { type: boolean }
...
    NewsItem:
      type: object
      required: [id, title, category, is_published, created_at]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: ["string", "null"] }
        category: { type: string }
        source: { type: ["string", "null"] }
        image_url: { type: ["string", "null"] }
        is_published: { type: boolean }
        published_at: { type: ["string", "null"] }
        created_at: { type: string }
    NewsInput:
      type: object
      required: [title, category]
      properties:
        title: { type: string }
        body: { type: string }
        category: { type: string }
        source: { type: string }
        image_url: { type: string }
        is_published: { type: boolean }
    NewsUpdate:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        category: { type: string }
        source: { type: string }
        image_url: { type: string }
        is_published: { type: boolean }
...
    PublicEvent:
      type: object
      required: [id, title, event_date, category, is_active, created_at]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        event_date: { type: string }
        category: { type: string }
        is_active: { type: boolean }
        created_at: { type: string }
    PublicEventInput:
      type: object
      required: [title, event_date, category]
      properties:
        title: { type: string }
        description: { type: string }
        event_date: { type: string }
        category: { type: string }
        is_active: { type: boolean }
    PublicEventUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        event_date: { type: string }
        category: { type: string }
        is_active: { type: boolean }
...